p43 - added "const" for some vec3_t arguments - game.dll support functions: added dummy functions for demo playback
git-svn-id: svn://localhost@54 2093b545-fd6f-4e46-8590-e32825573395

+++ release/fixvcdep.sh
@@ -4,15 +4,18 @@ tmpfile="tmp.dep"
 
 function FixDep()
 {
-	grep --invert-match --regexp=basetsd.h $1 > $tmpfile
-	mv -f $tmpfile $1
+	if [ -f $1 ]; then
+		grep --invert-match --regexp=basetsd.h $1 > $tmpfile
+		mv -f $tmpfile $1
+#		echo "parse $1"
+	fi
 }
 
 
 # fix main dependencies
 cd ..
 FixDep quake2s.dep
-#??FixDep quake2.dep
+FixDep quake2.dep
 
 # fix ref_gl dependencies
 cd ref_gl.old
